+++ test/HCurve.test.js
@@ -138,9 +138,9 @@ contract('Curve', function([_, user]) {
           token0User
         );
         // get_dy_underlying flow is different from exchange_underlying,
-        // so give 1*10^12 tolerance for USDT/DAI case.
+        // so give 2*10^12 tolerance for USDT/DAI case.
         expect(await this.token1.balanceOf.call(user)).to.be.bignumber.gte(
-          token1User.add(answer).sub(new BN('1000000000000'))
+          token1User.add(answer).sub(new BN('2000000000000'))
         );
         expect(await this.token1.balanceOf.call(user)).to.be.bignumber.lte(
           token1User.add(answer)
@@ -184,9 +184,9 @@ contract('Curve', function([_, user]) {
           token0User
         );
         // get_dy_underlying flow is different from exchange_underlying,
-        // so give 1*10^12 tolerance for USDT/DAI case.
+        // so give 2*10^12 tolerance for USDT/DAI case.
         expect(await this.token1.balanceOf.call(user)).to.be.bignumber.gte(
-          token1User.add(answer).sub(new BN('1000000000000'))
+          token1User.add(answer).sub(new BN('2000000000000'))
         );
         expect(await this.token1.balanceOf.call(user)).to.be.bignumber.lte(
           token1User.add(answer)
@@ -810,6 +810,74 @@ contract('Curve', function([_, user]) {
           this.poolToken.address,
           tokens,
           amounts,
+          minMintAmount
+        );
+        const receipt = await this.proxy.execMock(this.hCurve.address, data, {
+          from: user,
+          value: value,
+        });
+
+        // Get handler return
+        const handlerReturn = utils.toBN(
+          getHandlerReturn(receipt, ['uint256'])[0]
+        );
+        const poolTokenUserEnd = await this.poolToken.balanceOf.call(user);
+        expect(handlerReturn).to.be.bignumber.eq(
+          poolTokenUserEnd.sub(poolTokenUser)
+        );
+
+        // Check proxy balance
+        expect(await balanceProxy.get()).to.be.zero;
+        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.poolToken.balanceOf.call(this.proxy.address)
+        ).to.be.zero;
+
+        // Check user balance
+        expect(await balanceUser.delta()).to.be.bignumber.eq(
+          ether('0')
+            .sub(value)
+            .sub(new BN(receipt.receipt.gasUsed))
+        );
+        expect(await this.token.balanceOf.call(user)).to.be.bignumber.eq(
+          tokenUser
+        );
+
+        // poolToken amount should be greater than answer * 0.999 which is
+        // referenced from tests in curve contract.
+        expect(await this.poolToken.balanceOf.call(user)).to.be.bignumber.gte(
+          answer.mul(new BN('999')).div(new BN('1000'))
+        );
+        expect(await this.poolToken.balanceOf.call(user)).to.be.bignumber.lte(
+          answer
+        );
+
+        profileGas(receipt);
+      });
+
+      it('add ETH and sETH to pool by addLiquidity with max amount', async function() {
+        const value = ether('1');
+        const tokenAmount = ether('2');
+        const tokens = [ETH_TOKEN, this.token.address];
+        const amounts = [value, tokenAmount];
+
+        // Get expected answer
+        const answer = await this.sethSwap.methods[
+          'calc_token_amount(uint256[2],bool)'
+        ](amounts, true);
+
+        // Execute handler
+        await this.token.transfer(this.proxy.address, tokenAmount, {
+          from: providerAddress,
+        });
+        await this.proxy.updateTokenMock(this.token.address);
+        const minMintAmount = mulPercent(answer, new BN('100').sub(slippage));
+        const data = abi.simpleEncode(
+          'addLiquidity(address,address,address[],uint256[],uint256)',
+          this.sethSwap.address,
+          this.poolToken.address,
+          tokens,
+          [MAX_UINT256, MAX_UINT256],
           minMintAmount
         );
         const receipt = await this.proxy.execMock(this.hCurve.address, data, {